Simplify the calculations.
Instead of the complicated relative calculations, where I remembered the
last cursor pos and calculated deltas to add to the rectangle, now I
calculate the absolute sides of the new rectangle. All I needed to do
was to store the cursor distance from the 1 or 2 rectangle sides which
would be resizing, and then use that to calculate how the rectangle
should look for each cursor pos.

I got rid of lastCursorPos that way.